More Integration Tests
This commit adds unit tests and integration tests for the server binary.
All unit tests were moved to separate files to separate them from other
codes. Server unit tests are placeholders as endpoints still need to be
implemented.

The AddressSanitizer testing was changed to run without the asan feature
as the particular test now detects that test is run with root and
ignores the problematic part of the test.
