+++ .github/workflows/mobile_api-ubuntu.yml
@@ -229,7 +229,7 @@ jobs:
       env:
         RUSTFLAGS: -Zsanitizer=address
         RUSTDOCFLAGS: -Zsanitizer=address
-      run: cargo test -Zbuild-std --target x86_64-unknown-linux-gnu --features=asan
+      run: cargo test -Zbuild-std --target x86_64-unknown-linux-gnu
       # Use `cargo run` for the analysis of a binary.
       # Usage of the `help` command as base command, please replace it
       # with the effective command that AddressSanitizer has to analyze
