+++ create_device_info/Cargo.toml
@@ -14,10 +14,8 @@ qrcodegen = { version =  "1.8.0" }
 [dev-dependencies]
 assert_cmd = { version = "2.0.6" }
 image = { version = "0.24.5" }
+users = { version = "0.11.0" }
 predicates = { version = "2.1.3" }
 resvg = { version = "0.27.0" }
 rqrr = { version = "0.5.2" }
 tempfile = { version = "3.3.0" }
-
-[features]
-asan = [] # For disabling unit test when running AddressSanitizer
