+++ mobile_api_server/src/api_v1.rs
@@ -5,6 +5,9 @@ use rocket_okapi::openapi_get_routes;
 pub mod commands;
 pub mod device;
 
+#[cfg(test)]
+pub mod tests_common;
+
 /// Routes for the API v1
 ///
 /// Routes are run through [openapi_get_routes!] to generate OpenAPI specifications from
